requirements.txt
google-api-python-client==2.160.0
oauth2client==4.1.3
bidict==0.23.1
gspread==6.1.4
unidecode==1.3.8
pillow==11.1.0
qrcode==8.0

